Add random get/index get too.

+++ examples/benchy.py
@@ -1,6 +1,7 @@
 
 from __future__ import absolute_import
 import csv
+import gzip
 import operator
 import os
 import random
@@ -33,12 +34,13 @@ class CentiEngine(object):
         self.coll = self.store.collection('stuff',
             encoder=self.ENCODER, key_func=self.KEY_FUNC)
         if use_indices:
-            self.coll.add_index('rev_name', lambda p: p['name'][::-1])
-            self.coll.add_index('rev_locn', lambda p: p['location'][::-1])
+            self.coll.add_index('rev_name', lambda p: p['name'])
+            self.coll.add_index('rev_locn', lambda p: p['location'])
 
     def close(self):
         self.store.engine.close()
-        shutil.rmtree(self.PATH)
+        if self.PATH:
+            shutil.rmtree(self.PATH)
 
     def insert(self, words, upper, stub, blind):
         coll = self.coll
@@ -48,12 +50,25 @@ class CentiEngine(object):
             coll.put(doc, blind=blind, txn=txn)
         txn.commit()
 
+    def randget_idx(self, words):
+        index = self.coll.indices['rev_name']
+        txn = self.store.engine.begin()
+        for word in words:
+            index.get(word, txn=txn)
+
+    def randget_id(self, words, upper):
+        txn = self.store.engine.begin()
+        coll = self.coll
+        for i in xrange(len(words)):
+            coll.get((words[i], upper[i]), txn=txn)
+
 
 class LmdbEngine(CentiEngine):
     PATH = BASE_PATH + 'test.lmdb'
     def make_engine(self):
         self.store = centidb.open('LmdbEngine',
-            path=self.PATH, map_size=1048576*1024)
+            path=self.PATH, map_size=1048576*1024,
+            writemap=True)
 
 
 class SkiplistEngine(CentiEngine):
@@ -87,35 +102,70 @@ class MongoEngine(object):
     def insert(self, words, upper, stub, blind):
         coll = self.coll
         for i in xrange(len(words)):
-            doc = {'stub': stub, 'name': words[i], 'location': upper[i]}
+            doc = {'stub': stub, 'name': words[i], 'location': upper[i],
+                   '_id': '%s-%s' % (words[i], upper[i])}
             coll.insert(doc)
+
+    def randget_idx(self, words):
+        coll = self.coll
+        for word in words:
+            coll.find_one({'name': word})
+
+    def randget_id(self, words, upper):
+        coll = self.coll
+        for i in xrange(len(words)):
+            coll.find_one('%s-%s' % (words[i], upper[i]))
+
+    def close(self):
+        pass
 
 
 def x():
     global store
 
-    words = map(str.strip, file('/usr/share/dict/words'))
+    words = map(str.strip, gzip.open('words.gz'))
     random.shuffle(words)
     upper = map(str.upper, words)
     stub = len('%x' % (random.getrandbits(150*4)))
 
+    ids = range(len(words))
+    random.shuffle(ids)
+
     out('Engine', 'Blind', 'UseIndices', 'Keys', 'Time', 'Ops/s')
+    eng = None
     for engine in LmdbEngine, SkiplistEngine, PlyvelEngine, MongoEngine:
+        print
         for blind in True, False:
-            print
-            for use_indices in True, False:
+            for use_indices in False, True:
+                if eng:
+                    eng.close()
                 eng = engine()
                 eng.create()
                 eng.make_coll(use_indices)
 
                 t0 = time.time()
                 eng.insert(words, upper, stub, blind)
-                eng.close()
                 t = time.time() - t0
 
                 keycnt = len(words) * (3 if use_indices else 1)
                 engine_name = engine.__name__
                 out(engine_name, blind, use_indices, keycnt, '%.2f' % t,
-                    '%.2f' % ((keycnt if blind else (keycnt + len(words))) / t))
+                    int((keycnt if blind else (keycnt + len(words))) / t))
+
+        idxcnt = 0
+        t0 = time.time()
+        while time.time() < (t0 + 5):
+            eng.randget_idx(words)
+            idxcnt += len(words)
+        idxtime = time.time() - t0
+        out(engine_name, '-', 'idx', idxcnt, '%.2f' % t, int(idxcnt/idxtime))
+
+        idcnt = 0
+        t0 = time.time()
+        while time.time() < (t0 + 5):
+            eng.randget_id(words, upper)
+            idcnt += len(words)
+        idtime = time.time() - t0
+        out(engine_name, '-', 'id', idcnt, '%.2f' % t, int(idcnt/idtime))
 
 x()
